+++ include/overworld/Engine/Physics/PhysX/World.h
@@ -56,8 +56,8 @@ namespace owds::physx {
     std::unordered_set<int> getOverlappingActors(::owds::Actor* actor) override;
 
   private:
-    void performRaycastsInParallel(const std::vector<std::array<float, 3>>& origins,
-                                   const std::vector<std::array<float, 3>>& destinations,
+    void performRaycastsInParallel(const std::vector<std::array<double, 3>>& origins,
+                                   const std::vector<std::array<double, 3>>& destinations,
                                    float max_distance,
                                    std::vector<RaycastHitResult_t>& results) override;
   };
